chore: use github actions to deploy github pages.
Removes the usage of the chabad360/hugo-gh-pages Github action.
Replacing it with the official Github action to deploy Github pages.

This change means that we will stop deploy the Github pages in a branch
and start to use Github environments for that. Which may requires some
changes in the repository.

Signed-off-by: José Guilherme Vanz <jguilhermevanz@suse.com>